Fix invalid signatures
Problem: I thought `Uint8Array.from()` had the same API as
`Buffer.from()` but I was terribly incorrect. This was making a new
array with the same number of bytes, but they were all empty bytes,
which broke all of the signatures.

Solution: `new TextEncoder().encode()`, which converts a string to a
Uint8Array the way that we need it to.